Updated card structure and color scheme

+++ PokedexApp.Shared/Models/Pokemons/Pokemon.cs
@@ -11,7 +11,7 @@ public class Pokemon
 {
 	public int Id { get; set; }
 	public string Name { get; set; } = string.Empty;
-	public int Height { get; set; }
+	public double Height { get; set; }
 	public int Weight { get; set; }
 	public PokemonCries Cries{get; set;}
 	public PokemonSprites Sprites { get; set; }
